requirements.txt
absl-py==2.3.1
asgiref==3.9.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Django==5.2.5
djangorestframework==3.16.1
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==23.0.0
h5py==3.14.0
huggingface-hub==0.15.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
keras==2.13.1
libclang==18.1.1
Markdown==3.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.24.3
oauthlib==3.3.1
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.1
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.33.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
